requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.1.0=gpu
absl-py=0.9.0=py37_0
aiofiles=0.5.0=py_0
astor=0.8.0=py37_0
backcall=0.1.0=py37_0
blas=1.0=mkl
blinker=1.4=py37_0
ca-certificates=2020.6.20=hecda079_0
cachetools=3.1.1=py_0
certifi=2020.6.20=py37hc8dfbb8_0
cffi=1.14.0=py37h7a1dbc1_0
chardet=3.0.4=py37_1003
click=7.1.2=py_0
colorama=0.4.3=py_0
cryptography=2.9.2=py37h7a1dbc1_0
cudatoolkit=10.1.243=h74a9793_0
cudnn=7.6.5=cuda10.1_0
cycler=0.10.0=py37_0
decorator=4.4.2=py_0
entrypoints=0.3=py37_0
fastapi=0.60.0=py_0
flask=1.1.2=py_0
flask-cors=3.0.8=py_0
freetype=2.9.1=ha9979f8_1
gast=0.2.2=py37_0
google-api-core=1.20.0=py37_0
google-api-python-client=1.10.0=pyh9f0ad1d_0
google-auth=1.19.2=py_0
google-auth-httplib2=0.0.3=py_3
google-auth-oauthlib=0.4.1=py_2
google-pasta=0.2.0=py_0
googleapis-common-protos=1.51.0=py37h1f41ff6_2
graphene=1.4=py_0
graphql-core=3.1.2=py_0
graphql-relay=3.0.0=py_0
graphviz=2.38=hfd603c8_2
grpcio=1.27.2=py37h351948d_0
h11=0.9.0=py_0
h5py=2.8.0=py37hf7173ca_2
hdf5=1.8.20=hac2f561_1
httplib2=0.18.1=pyh9f0ad1d_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=vc14hc45fdbb_0
idna=2.9=py_1
imutils=0.5.3=py37_1
intel-openmp=2020.1=216
ipykernel=5.1.4=py37h39e3cac_0
ipython=7.13.0=py37h5ca1d4c_0
ipython_genutils=0.2.0=py37_0
itsdangerous=1.1.0=py_0
jedi=0.17.0=py37_0
jinja2=2.11.2=py_0
joblib=0.15.1=py_0
jpeg=9b=vc14h4d7706e_1
jupyter_client=6.1.3=py_0
jupyter_core=4.6.3=py37_0
keras-applications=1.0.8=py_0
keras-preprocessing=1.1.0=py_1
kiwisolver=1.2.0=py37h74a9793_0
libopencv=3.4.2=h20b85fd_0
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.11.4=h7bd577a_0
libsodium=1.0.16=h9d3ae62_0
libtiff=4.1.0=h56a325e_1
lz4-c=1.9.2=h62dcd97_0
markdown=3.1.1=py37_0
markupsafe=1.1.1=py37he774522_0
matplotlib=3.1.2=py37_1
matplotlib-base=3.1.2=py37h64f37c6_1
mkl=2020.1=216
mkl-service=2.3.0=py37hb782905_0
mkl_fft=1.0.15=py37h14836fe_0
mkl_random=1.1.1=py37h47e9c7a_0
numpy=1.18.1=py37h93ca92e_0
numpy-base=1.18.1=py37hc3f5095_1
oauth2client=4.1.3=py_0
oauthlib=3.1.0=py_0
opencv=3.4.2=py37h40b0b35_0
openssl=1.1.1g=he774522_1
opt_einsum=3.1.0=py_0
pandas=1.0.3=py37h47e9c7a_0
parso=0.7.0=py_0
pickleshare=0.7.5=py37_0
pip=20.0.2=py37_3
promise=2.3=py37hc8dfbb8_1
prompt-toolkit=3.0.5=py_0
prompt_toolkit=3.0.5=0
protobuf=3.11.4=py37h33f27b4_0
py-opencv=3.4.2=py37hc319ecb_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.20=py_0
pydantic=1.6.1=py37h4ab8f01_0
pydot=1.4.1=py37_0
pydrive=1.3.1=py_1
pygments=2.6.1=py_0
pyjwt=1.7.1=py37_0
pymongo=3.9.0=py37ha925a31_0
pyopenssl=19.1.0=py37_0
pyparsing=2.4.7=py_0
pyqt=5.9.2=py37h6538335_2
pyreadline=2.1=py37_1
pysocks=1.7.1=py37_0
python=3.7.6=h60c2a47_2
python-dateutil=2.8.1=py_0
python-multipart=0.0.5=py_0
python_abi=3.7=1_cp37m
pytz=2020.1=py_0
pywin32=227=py37he774522_1
pyyaml=5.3.1=py37h8055547_0
pyzmq=18.1.1=py37ha925a31_0
qt=5.9.7=vc14h73c81de_0
requests=2.23.0=py37_0
requests-oauthlib=1.3.0=py_0
rsa=4.0=py_0
scikit-learn=0.22.1=py37h6288b17_0
scipy=1.4.1=py37h9439919_0
setuptools=46.4.0=py37_0
simplejson=3.17.2=py37h4ab8f01_0
sip=4.19.8=py37h6538335_0
six=1.14.0=py37_0
sqlite=3.31.1=h2a8f88b_1
starlette=0.13.4=py_0
tensorboard=2.2.1=pyh532a8cf_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow=2.1.0=gpu_py37h7db9008_0
tensorflow-base=2.1.0=gpu_py37h55f5790_0
tensorflow-estimator=2.1.0=pyhd54b08b_0
tensorflow-gpu=2.1.0=h0d30ee6_0
termcolor=1.1.0=py37_1
tornado=6.0.4=py37he774522_1
traitlets=4.3.3=py37_0
ujson=3.0.0=py37h1834ac0_0
uritemplate=3.0.1=py_0
urllib3=1.25.8=py37_0
uvicorn=0.11.3=py37hc8dfbb8_1
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_2
wcwidth=0.1.9=py_0
websockets=8.1=py37h8055547_1
werkzeug=0.16.1=py_0
wheel=0.34.2=py37_0
win_inet_pton=1.1.0=py37_0
wincertstore=0.2=py37_0
wrapt=1.12.1=py37he774522_1
wsproto=0.13.0=py_0
xz=5.2.5=h62dcd97_0
yaml=0.2.5=he774522_0
zeromq=4.3.1=h33f27b4_3
zlib=1.2.11=vc14h1cdd9ab_1
zstd=1.4.4=ha9fde0e_3
